NEED HELP WITH SORT

+++ checkpointbuglogger.client/src/components/note.vue
@@ -1,13 +1,16 @@
 <template>
-  <div class="note row border-right">
-    <div class="col-md-3 mx-5 border-right border-bottom">
-      <b>Creator Name</b>
-    </div>
-    <div class="col-md-9 mx-5 border-right border-bottom">
-      <b>{{ noteProp.body }}</b>
-    </div>
-    <div>
-      <i class="fas fa-trash-alt    " @click="deleteNote"></i>
+  <div class="container-fluid">
+    <div class="note row border-right">
+      <div class="col-md-3 mx-5 border-right border-bottom">
+        <!--WHY WON'T YOUR CREATOR POPULATE-->
+        <b>Creator Name</b>
+      </div>
+      <div class="col-md-9 mx-5 border-right border-bottom">
+        <b>{{ noteProp.body }}</b>
+      </div>
+      <div>
+        <i class="fas fa-trash-alt    " @click="deleteNote" v-if="state.user.isAuthenticated && state.account.id === noteProp.creatorId"></i>
+      </div>
     </div>
   </div>
 </template>
@@ -30,14 +33,16 @@ export default {
       account: computed(() => AppState.account),
       user: computed(() => AppState.user)
     })
+
     return {
       state,
       async deleteNote() {
         try {
-          await notesService.deleteNote(props.noteProp.id, props.noteProp.bug)
-          Notification.toast('Successfully Deleted Your Note', 'success')
+          if (await Notification.confirmAction()) {
+            await notesService.deleteNote(props.noteProp.id, props.noteProp.bug)
+          }
         } catch (error) {
-          Notification.toast('Error: ' + error, 'warning')
+          Notification.toast('Error: ' + error, 'error')
         }
       }
     }
